Changing radius property from Float to CGFloat

+++ Code/Views/LYRUIProgressView.m
@@ -43,7 +43,7 @@ NSString *LYRUIProgressViewStringForIconStyle(LYRUIProgressViewIconStyle iconSty
 @property (nonatomic) CATextLayer *iconLayer;
 @property (nonatomic) UIBezierPath *progressArcPath;
 @property (nonatomic) CGFloat borderWidth;
-@property (nonatomic, readonly) float radius;
+@property (nonatomic, readonly) double radius;
 @property (nonatomic) NSTimeInterval animationDuration;
 @property (nonatomic) double progress;
